test: refactor struct array element tests — split search/query, unify naming and style
- Split test_struct_array_element.py into query and search files
- Move TestGroupBySearch + TestElementLevelSearchNoFilter to search file
- Move TestStructArrayMatchQuery to query file
- Unify class names to TestMilvusClient* convention (20 classes)
- Unify test method prefixes for consistency within each class
- Use ct.default_nb, ct.default_dim, add common.constants import
- Fix silent no-op range filter test (10,50) -> (100,200)
- Remove dead code: gt_element_filter_search and unused helpers
- Remove unused constants: TAG_POOL, SIZES
- Move appended constants to file header in search file
